move NodeIdentity to more suitable package
